Proper way to handle permissions
